初始化

+++ src/java/com/bootdo/gc/service/impl/KehuServiceImpl.java
@@ -1,13 +1,19 @@
 package com.bootdo.gc.service.impl;
 
+import com.bootdo.common.utils.DateUtils;
 import com.bootdo.gc.dao.KehuDao;
 import com.bootdo.gc.domain.KehuDO;
 import com.bootdo.gc.domain.SijiDO;
+import com.bootdo.gc.domain.SijiItemDO;
 import com.bootdo.gc.service.KehuService;
+import com.bootdo.gc.service.SijiItemService;
 import com.bootdo.gc.service.SijiService;
+import org.apache.commons.collections.map.HashedMap;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
+import java.math.BigDecimal;
+import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
 
@@ -19,6 +25,10 @@ public class KehuServiceImpl implements KehuService {
 	private KehuDao kehuDao;
 
 	@Autowired
+	private SijiItemService sijiItemService;
+
+
+	@Autowired
 	private SijiService sijiService;
 	
 	@Override
@@ -28,17 +38,43 @@ public class KehuServiceImpl implements KehuService {
 
 	@Override
 	public KehuDO getOrder(String orderNo){
-		KehuDO kehuDO = new KehuDO();
 
+		//存在此单直接返回
+		KehuDO kehu = kehuDao.getOrder(orderNo);
+		if (kehu!=null) {
+			return kehu;
+		}
+
+		KehuDO kehuDO = new KehuDO();
 		SijiDO sijiDO = sijiService.getOrder(orderNo);
-		if (kehuDO==null) {
-			return null;
+		if (sijiDO==null) {
+			kehuDO.setOrderNo(orderNo);
+			return kehuDO;
 		}
 		kehuDO.setBizdate(sijiDO.getBizdate());
 		kehuDO.setCarnum(sijiDO.getCarnum());
 		kehuDO.setForwunit(sijiDO.getForwunit());
 		kehuDO.setStation(sijiDO.getArrivstation());
-		//kehuDO.setTranscost(sijiDO.getTrancost());
+		kehuDO.setKpid(sijiDO.getPid());
+
+		Map map = new HashMap();
+		map.put("pid",sijiDO.getPid());
+		kehuDO.setTonnage(new BigDecimal(0));
+		List<SijiItemDO> list = sijiItemService.list(map);
+		for (SijiItemDO sijiItemDO : list){
+			kehuDO.setTonnage(kehuDO.getTonnage().add(sijiItemDO.getTonnage())); //吨位
+		}
+
+
+		try {
+			if (kehuDO.getTonnage()!=null){
+				kehuDO.setSettletonnage(kehuDO.getTonnage().setScale(1, BigDecimal.ROUND_DOWN)); //结算吨位
+			}
+			//kehuDO.setAdjusttonnage(sijiDO.getTonnage()); //调整吨位
+		} catch (Exception e){
+
+		}
+
 
 
 		return kehuDO;
@@ -67,12 +103,40 @@ public class KehuServiceImpl implements KehuService {
 	
 	@Override
 	public int remove(Long id){
-		return kehuDao.remove(id);
+
+		KehuDO kehuDO = kehuDao.get(id);
+		boolean flag = "有".equals(kehuDO.getReceipt())||"预有".equals(kehuDO.getReceipt());
+		if (flag && !DateUtils.getNowDate().equals(kehuDO.getReceiptdate())) {
+			return 0;
+		} else {
+			return kehuDao.remove(id);
+		}
 	}
 	
 	@Override
 	public int batchRemove(Long[] ids){
+		for (Long id :ids){
+			KehuDO kehuDO = kehuDao.get(id);
+			boolean flag = "有".equals(kehuDO.getReceipt())||"预有".equals(kehuDO.getReceipt());
+			if (flag && !DateUtils.getNowDate().equals(kehuDO.getReceiptdate())) {
+				return 0;
+			}
+		}
+
 		return kehuDao.batchRemove(ids);
+	}
+
+
+	@Override
+	public List<Map> queryKehu(Map map) {
+
+		return   kehuDao.queryKehu(map);
+	}
+
+	@Override
+	public List<Map> queryKehu1(Map map) {
+
+		return   kehuDao.queryKehu1(map);
 	}
 	
 }
